Fixing font and modularizing components
