+++ src/styles/components/_index.scss
@@ -2,3 +2,5 @@
 @import 'card';
 @import 'input';
 @import 'navbar';
+@import 'backdrop';
+@import 'drawer'
